Rename hctr2-xctr to xctr
